+++ src/components/birdSightings/BirdSightingsTable.scss
