+++ src/components/birds/BirdCard.scss
