+++ src/components/birds/BirdDetail.scss
