+++ src/components/birds/BirdFeatured.js
