+++ src/components/birds/BirdFeatured.scss
