+++ src/components/birds/BirdProfile.scss
