+++ src/components/sightings/SightingDetail.js
