+++ src/components/sightings/SightingDetail.scss
