+++ src/components/sightings/SightingsMap.js
