+++ src/components/sightings/SightingsTable.scss
