+++ src/views/birds/index.js
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import Helmet from 'react-helmet';
 
 import Banner from '../../components/presentation/Banner/Banner';
-import Birds from '../../components/Bird/Birds';
-// import BirdsFilterForm from '../../components/Bird/search/BirdsFilterForm';
-import BirdsSearchForm from '../../components/Bird/search/BirdsSearchForm';
+import Birds from '../../components/birds/Birds';
+// import BirdsFilterForm from '../../components/birds/search/BirdsFilterForm';
+import BirdsSearchForm from '../../components/birds/search/BirdsSearchForm';
 import { fetchBandCombosIfNeeded } from '../../actions/bandCombos.js';
 
 class BirdsPage extends Component {
