+++ src/views/sightings/detail.js
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import Helmet from 'react-helmet';
 
 import Banner from '../../components/presentation/Banner/Banner';
-import SightingDetail from '../../components/Sightings/SightingDetail';
-import BirdSightingsTable from '../../components/BirdSightings/BirdSightingsTable';
-import SightingsMap from '../../components/Sightings/SightingsMap';
+import SightingDetail from '../../components/sightings/SightingDetail';
+import BirdSightingsTable from '../../components/birdSightings/BirdSightingsTable';
+import SightingsMap from '../../components/sightings/SightingsMap';
 
 class SightingDetailPage extends Component {
   render() {
